Efficiency Improvements & Bug Fixes
Improved fetching efficiency so that is only 2-core calls instead of 3-4. Note that for larger desks, the fetch may take some time due to the number of fetches required for a full month - Zoho limits to 100 tickets per full, with no way to reduce content received on output. So if there is more than 100, need to loop the fetch and page through to get each set of 100 ticket data.
This may also need a lot of RAM - Am looking into better ways to do this - please let me know if you know of better ways too.

Added Ticket Status filter input.

Increased the refresh time to 120 seconds, instead of 60 seconds.

Made some adjustments to the UI. Might make more later, not happy with yet.

A few bug fixes, and removed a bunch of extra code that is no longer required.